# app/resources/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOTR API",
        "description": "Description of LOTR API",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "url",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/characters": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Retrieve a paginated list of characters",
                "operationId": "d15bf9b69dba07755d4c31b6c8944414",
                "responses": {
                    "200": {
                        "description": "A paginated list of characters returned successfully."
                    }
                }
            },
            "post": {
                "tags": [
                    "Characters"
                ],
                "summary": "Store a newly created character",
                "operationId": "ed765ff69503e7e25318df57eca145a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CharacterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Character created successfully."
                    },
                    "400": {
                        "description": "Unable to create character."
                    }
                }
            }
        },
        "/api/v1/characters/{id}": {
            "get": {
                "tags": [
                    "Characters"
                ],
                "summary": "Retrieve a specific character by ID",
                "operationId": "f9360ec9f5850d7ba238b2656a9f0202",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the character to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character retrieved successfully."
                    },
                    "404": {
                        "description": "Character not found."
                    }
                }
            },
            "put": {
                "tags": [
                    "Characters"
                ],
                "summary": "Update a character",
                "operationId": "8a25ed967a79e3b9d087c33626129c23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the character to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CharacterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Character updated successfully."
                    },
                    "404": {
                        "description": "Character not found."
                    },
                    "400": {
                        "description": "Unable to update character."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Characters"
                ],
                "summary": "Soft delete a character",
                "operationId": "b6b554f97083f12ac30995d9eebdf617",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the character to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Character deleted successfully."
                    },
                    "404": {
                        "description": "Character not found."
                    },
                    "400": {
                        "description": "Unable to delete character."
                    }
                }
            }
        },
        "/api/v1/characters/{id}/restore": {
            "post": {
                "tags": [
                    "Characters"
                ],
                "summary": "Restore a soft-deleted character",
                "operationId": "ea5073cc345109006c89b4e51c95147f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the character to restore",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character restored successfully."
                    },
                    "404": {
                        "description": "Character not found."
                    },
                    "400": {
                        "description": "Unable to restore character."
                    }
                }
            }
        },
        "/api/v1/characters/{id}/force-delete": {
            "delete": {
                "tags": [
                    "Characters"
                ],
                "summary": "Permanently delete a character",
                "operationId": "ec5fdc14515c593180875bba087afb6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the character to permanently delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Character permanently deleted successfully."
                    },
                    "404": {
                        "description": "Character not found."
                    },
                    "400": {
                        "description": "Unable to permanently delete character."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CharacterRequest": {
                "required": [
                    "name",
                    "birth_date",
                    "kingdom",
                    "equipment_id",
                    "faction_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 128,
                        "example": "Aragorn"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1980-01-15"
                    },
                    "kingdom": {
                        "type": "string",
                        "maxLength": 128,
                        "example": "Gondor"
                    },
                    "equipment_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "faction_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}